chore: remove code of conduct
I've recently realised that I was strongarmed into adding a CoC, a concept I don't give any backing to in the context of repositories and their environment, to my repository. GitHub seems to imply that your repo is incomplete without a CoC, much like how missing a README makes it incomplete.

After reading and thinking about a few articles and emails from Stallman, I've decided to remove the CoC from paste, not that anyone will really notice or care. :)

[skip ci]