feat: add JS and no-JS "captcha" options
JS is done in the background and basically just confirms the browser has working JavaScript. This is just an attempt to stop basic bots from interacting with the submission form.

The no-JS option is handled by asking the submitter to answer a simple addition problem, which, again, is just to stop simple bots that aren't customised for paste.